gcs_sqlite_to_bq/requirements.txt
pycountry
SQLAlchemy==1.3.22
spacy
pandas
google-cloud-storage
google-cloud-datastore
google-cloud-bigquery
